refactor: organize pages by feature folders
