+++ src/App.tsx
@@ -14,14 +14,14 @@ import { ErrorBoundary } from "@/components/ErrorBoundary";
 // Pages
 import Home from "./pages/Home";
 import Dashboard from "./pages/Dashboard";
-import RegisterAsset from "./pages/RegisterAsset";
-import AssetsDashboard from "./pages/AssetsDashboard";
-import AssetsInventory from "./pages/AssetsInventory";
-import AssetsManagement from "./pages/AssetsManagement";
-import AssetAssociation from "./pages/AssetAssociation";
-import AssociationsList from "./pages/AssociationsList";
-import Clients from "./pages/Clients";
-import Association from "./pages/Association";
+import RegisterAsset from "./pages/assets/Register";
+import AssetsDashboard from "./pages/assets/Dashboard";
+import AssetsInventory from "./pages/assets/Inventory";
+import AssetsManagement from "./pages/assets/Management";
+import AssetAssociation from "./pages/associations/Asset";
+import AssociationsList from "./pages/associations/List";
+import Clients from "./pages/clients/List";
+import Association from "./pages/associations/Association";
 import Subscriptions from "./pages/Subscriptions";
 import Monitoring from "./pages/Monitoring";
 import History from "./pages/History";
