+++ src/pages/assets/Inventory.tsx
