+++ src/pages/assets/index.ts
@@ -1,0 +1,6 @@
+export { default as AddAsset } from './Add';
+export { default as AssetHistory } from './History';
+export { default as AssetsDashboard } from './Dashboard';
+export { default as AssetsInventory } from './Inventory';
+export { default as AssetsManagement } from './Management';
+export { default as RegisterAsset } from './Register';
