+++ src/pages/associations/index.ts
@@ -1,0 +1,3 @@
+export { default as AssetAssociation } from './Asset';
+export { default as AssociationsList } from './List';
+export { default as Association } from './Association';
