+++ src/pages/clients/index.ts
@@ -1,0 +1,1 @@
+export { default as Clients } from './List';
